+++ CHANGES.md
@@ -1,5 +1,21 @@
 # Hollo changelog
 
+## Version 0.3.0
+
+To be released.
+
+- Added support for local filesystem storage for media files.
+  You can now configure `DRIVE_DISK=fs` and `FS_ASSET_PATH` to store media
+  files in the local filesystem. [[#59]]
+
+  - Added `DRIVE_DISK` environment variable.
+  - Added `FS_ASSET_PATH` environment variable.
+  - Added `ASSET_URL_BASE` environment variable to replace `S3_URL_BASE`.
+  - Deprecated `S3_URL_BASE` environment variable in favor of
+    `ASSET_URL_BASE`.
+
+[#59]: https://github.com/dahlia/hollo/pull/59
+
 ## Version 0.2.1
 
 Released on November 4, 2024.